deps/v8/test/inspector/debugger/side-effect-free-debug-evaluate-expected.txt

Tests side-effect-free evaluation

Running test: basicTest
Paused on "debugger;"
f() returns 1
g() returns 2
f() returns 1
g() throws EvalError

Running test: testAsyncFunctions
testAsyncFunction("resolve") : ok
testAsyncFunction("reject") : throws

Running test: testDate
someGlobalDate.setDate(10) : throws
new Date().setDate(10) : ok
someGlobalDate.setFullYear(1991) : throws
new Date().setFullYear(1991) : ok
someGlobalDate.setHours(0) : throws
new Date().setHours(0) : ok
someGlobalDate.getDate() : ok
new Date().getDate() : ok
someGlobalDate.getFullYear() : ok
new Date().getFullYear() : ok
someGlobalDate.getHours() : ok
new Date().getHours() : ok

Running test: testPromiseReject
Promise.reject() : throws

Running test: testSpread
[...someGlobalArray] : ok
[...someGlobalArray.values()] : ok
[...someGlobalArrayIterator] : throws
[...someGlobalMap] : ok
[...someGlobalMap.keys()] : ok
[...someGlobalMap.values()] : ok
[...someGlobalMapKeysIterator] : throws
[...someGlobalMapValuesIterator] : throws
[...someGlobalSet] : ok
[...someGlobalSet.values()] : ok
[...someGlobalSetIterator] : throws
